+++ shiksha-website/shiksha-backend/dao/master.subject.dao.js
@@ -1,9 +1,66 @@
 const MasterSubject = require("../models/master.subject.model.js");
 const BaseDao = require("./base.dao.js");
+const mongoose = require("mongoose");
+
+// --- Helpers ---
+const regexExact = (val) => new RegExp(`^${String(val).trim()}$`, "i");
+const str = (val) => String(val || "").trim();
 
 class MasterSubjectDao extends BaseDao {
 	constructor() {
 		super(MasterSubject);
+	}
+
+	async resolveSubjectName(identifier) {
+		if (mongoose.Types.ObjectId.isValid(identifier)) {
+			const subjectDoc = await MasterSubject.findById(identifier)
+				.select("name")
+				.lean();
+			return subjectDoc ? subjectDoc.name : str(identifier);
+		}
+		const subjectDoc = await MasterSubject.findOne({
+			$or: [{ name: identifier }, { subjectName: regexExact(identifier) }],
+		})
+			.select("name")
+			.lean();
+
+		return subjectDoc ? subjectDoc.name : str(identifier);
+	}
+
+	async resolveSubjectContext(identifier) {
+		let subjectCode = str(identifier);
+		let targetSubjectIds = [];
+
+		if (mongoose.Types.ObjectId.isValid(identifier)) {
+			const subjectDoc = await MasterSubject.findById(identifier)
+				.select("name")
+				.lean();
+			if (subjectDoc) {
+				subjectCode = subjectDoc.name;
+				const relatedSubjects = await MasterSubject.find({ name: subjectCode })
+					.select("_id")
+					.lean();
+				targetSubjectIds = relatedSubjects.map((s) => s._id);
+			} else {
+				targetSubjectIds = [new mongoose.Types.ObjectId(identifier)];
+			}
+		} else {
+			const relatedSubjects = await MasterSubject.find({
+				$or: [
+					{ name: subjectCode },
+					{ subjectName: regexExact(subjectCode) },
+				],
+			})
+				.select("_id name")
+				.lean();
+
+			if (relatedSubjects.length > 0) {
+				targetSubjectIds = relatedSubjects.map((s) => s._id);
+				if (relatedSubjects[0].name) subjectCode = relatedSubjects[0].name;
+			}
+		}
+
+		return { subjectCode, targetSubjectIds };
 	}
 
 	async getByNameAndBoard(subjectName, board) {
